Add operations only permissions

+++ backend/contact_book/src/main/java/com/app/contact_book/controllers/ContactController.java
@@ -4,11 +4,13 @@ import static com.app.contact_book.utils.Utils.validation;
 
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -62,33 +64,41 @@ public class ContactController {
     }
 
     @PostMapping
-    public ResponseEntity<?> create(@Valid @RequestBody ContactDTO contact, BindingResult result){
+    public ResponseEntity<?> create(@Valid @RequestBody ContactDTO contact, BindingResult result,
+        @AuthenticationPrincipal String userId){
         if(result.hasErrors()){
             return validation(result);
         }
-        Optional<User> userOptional = this.userService.findById(contact.getUserId());
-        if(userOptional.isPresent()){
-            return ResponseEntity.status(HttpStatus.CREATED).body(this.contactService.save(contact));
-        }
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(Collections.singletonMap("Error", "User ID not found"));
+        contact.setUserId(Long.parseLong(userId));
+        return ResponseEntity.status(HttpStatus.CREATED).body(this.contactService.save(contact));
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> update(@Valid @RequestBody ContactUpdateDTO contact, BindingResult result, @PathVariable Long id){
+    public ResponseEntity<?> update(@Valid @RequestBody ContactUpdateDTO contact, BindingResult result, @PathVariable Long id,
+        @AuthenticationPrincipal String userId){
         if(result.hasErrors()){
             return validation(result);
         }
-        Optional<ContactDTO> contactOptional = this.contactService.update(id, contact);
-        if(contactOptional.isPresent()){
-            return ResponseEntity.ok(contactOptional.get());
+        Optional<ContactDTO> contactOp = this.contactService.findById(id);
+        if(contactOp.isPresent()){
+            if(!contactOp.get().getUserId().equals(Long.parseLong(userId))){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Map.of("error", "You don't have permission to modify this contact"));
+            }
+            Optional<ContactDTO> contactOptional = this.contactService.update(id, contact);
+            if(contactOptional.isPresent()){
+                return ResponseEntity.ok(contactOptional.get());
+            }
         }
         return ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> delete(@PathVariable Long id){
+    public ResponseEntity<?> delete(@PathVariable Long id, @AuthenticationPrincipal String userId){
         Optional<ContactDTO> contactOptional = this.contactService.findById(id);
         if(contactOptional.isPresent()){
+            if(!contactOptional.get().getUserId().equals(Long.parseLong(userId))){
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).body(Map.of("error", "You don't have permission to modify this contact"));
+            }
             this.contactService.deleteById(id);
             return ResponseEntity.noContent().build();
         }
